Springboot MVC Eshopper Schema oe add table product_image

+++ src/eshopper-backend/src/main/java/codeid/eshopper_backend/model/entity/ProductImage.java
@@ -29,7 +29,7 @@ public class ProductImage extends AbstractEntity{
     private String fileName;
 
     @Column(name="file_size")
-    private Double fileSize;
+    private Long fileSize;
 
     @Column(name="file_type")
     private String fileType;
